refactor: organiza as pastas da aplicação
