+++ index.html
@@ -8,7 +8,7 @@
 	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 	<link href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap" rel="stylesheet">
 
-	<link rel="stylesheet" href="./style.css">
+	<link rel="stylesheet" href="./styles/style.css">
 	
 	<title>Viginere</title>
 </head>
@@ -111,7 +111,7 @@
 		</div>
 	</footer>
 
-	<script src="./script.js" type="module"></script>
+	<script src="./scripts/script.js" type="module"></script>
 
 </body>
 </html>
