+++ utils/criptografar.js
